Crate literalext

This crate provides extension methods to proc-macro, and proc-macro2's Literal types. These methods provide a mechanism for extracting the value of the type.

Supported Features

i128: Add support for interpreting the i128 and u128 integer types. nightly only

proc-macro2 default: Implement LiteralExt on proc_macro2::Literal.

proc-macro: Implement LiteralExt on proc_macro::Literal. nightly only

dummy: Export a type DummyLiteral with a public constructor which implements the LiteralExt trait.

Structs

FloatLit
A type which represents a floating point value.

IntLit
A type which represents an integer literal.

Traits

LiteralExt