Struct literalext::IntLit

pub struct IntLit { /* fields omitted */ }

A type which represents an integer literal.

Methods

impl IntLit

pub fn suffix(&self) -> &str

Get the suffix written on the integer literal.

pub fn as_u8(&self) -> Option<u8>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_i8(&self) -> Option<i8>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_u16(&self) -> Option<u16>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_i16(&self) -> Option<i16>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_u32(&self) -> Option<u32>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_i32(&self) -> Option<i32>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_u64(&self) -> Option<u64>

Returns None if the value overflows, or if the suffix is wrong.

pub fn as_i64(&self) -> Option<i64>

Returns None if the value overflows, or if the suffix is wrong.